Format with black

+++ dask_ml/metrics/regression.py
@@ -12,18 +12,20 @@ def _check_sample_weight(sample_weight):
     if sample_weight is not None:
         no_average = DASK_VERSION <= packaging.version.parse("0.18.0")
         if no_average:
-            raise NotImplementedError("'sample_weight' is only supported for "
-                                      "dask versions > 0.18.0.")
+            raise NotImplementedError(
+                "'sample_weight' is only supported for " "dask versions > 0.18.0."
+            )
 
 
 def _check_reg_targets(y_true, y_pred, multioutput):
-    allowed_multioutput_str = ('raw_values', 'uniform_average',
-                               'variance_weighted')
+    allowed_multioutput_str = ("raw_values", "uniform_average", "variance_weighted")
     if isinstance(multioutput, six.string_types):
         if multioutput not in allowed_multioutput_str:
-            raise ValueError("Allowed 'multioutput' string values are {}. "
-                             "You provided multioutput={!r}"
-                             "".format(allowed_multioutput_str, multioutput))
+            raise ValueError(
+                "Allowed 'multioutput' string values are {}. "
+                "You provided multioutput={!r}"
+                "".format(allowed_multioutput_str, multioutput)
+            )
     if y_true.ndim == 1:
         y_true = y_true.reshape((-1, 1))
     if y_pred.ndim == 1:
@@ -34,21 +36,17 @@ def _check_reg_targets(y_true, y_pred, multioutput):
 
 
 @doc_wraps(sklearn.metrics.mean_squared_error)
-def mean_squared_error(y_true, y_pred,
-                       sample_weight=None,
-                       multioutput='uniform_average',
-                       compute=True):
+def mean_squared_error(
+    y_true, y_pred, sample_weight=None, multioutput="uniform_average", compute=True
+):
     _check_sample_weight(sample_weight)
-    _, y_true, y_pred, multioutput = _check_reg_targets(
-        y_true, y_pred, multioutput
-    )
-    output_errors = da.average((y_pred - y_true) ** 2, axis=0,
-                               weights=sample_weight)
+    _, y_true, y_pred, multioutput = _check_reg_targets(y_true, y_pred, multioutput)
+    output_errors = da.average((y_pred - y_true) ** 2, axis=0, weights=sample_weight)
 
     if isinstance(multioutput, six.string_types):
-        if multioutput == 'raw_values':
+        if multioutput == "raw_values":
             return output_errors.compute() if compute else output_errors
-        elif multioutput == 'uniform_average':
+        elif multioutput == "uniform_average":
             # pass None as weights to da.average: uniform mean
             multioutput = None
     result = da.average(output_errors, weights=multioutput)
@@ -59,21 +57,17 @@ def mean_squared_error(y_true, y_pred,
 
 
 @doc_wraps(sklearn.metrics.mean_squared_error)
-def mean_absolute_error(y_true, y_pred,
-                        sample_weight=None,
-                        multioutput='uniform_average',
-                        compute=True):
+def mean_absolute_error(
+    y_true, y_pred, sample_weight=None, multioutput="uniform_average", compute=True
+):
     _check_sample_weight(sample_weight)
-    _, y_true, y_pred, multioutput = _check_reg_targets(
-        y_true, y_pred, multioutput
-    )
-    output_errors = da.average(abs(y_pred - y_true), axis=0,
-                               weights=sample_weight)
+    _, y_true, y_pred, multioutput = _check_reg_targets(y_true, y_pred, multioutput)
+    output_errors = da.average(abs(y_pred - y_true), axis=0, weights=sample_weight)
 
     if isinstance(multioutput, six.string_types):
-        if multioutput == 'raw_values':
+        if multioutput == "raw_values":
             return output_errors.compute() if compute else output_errors
-        elif multioutput == 'uniform_average':
+        elif multioutput == "uniform_average":
             # pass None as weights to da.average: uniform mean
             multioutput = None
     result = da.average(output_errors, weights=multioutput)
@@ -83,13 +77,11 @@ def mean_absolute_error(y_true, y_pred,
 
 
 @doc_wraps(sklearn.metrics.r2_score)
-def r2_score(y_true, y_pred, sample_weight=None,
-             multioutput="uniform_average",
-             compute=True):
+def r2_score(
+    y_true, y_pred, sample_weight=None, multioutput="uniform_average", compute=True
+):
     _check_sample_weight(sample_weight)
-    _, y_true, y_pred, multioutput = _check_reg_targets(
-        y_true, y_pred, multioutput
-    )
+    _, y_true, y_pred, multioutput = _check_reg_targets(y_true, y_pred, multioutput)
     if sample_weight is not None:
         weight = sample_weight
         if weight.ndim == 1:
@@ -97,26 +89,26 @@ def r2_score(y_true, y_pred, sample_weight=None,
     else:
         weight = 1.0
 
-    numerator = (weight * (y_true - y_pred) ** 2).sum(axis=0, dtype='f8')
-    denominator = (weight * (y_true - da.average(
-        y_true, axis=0, weights=sample_weight)) ** 2).sum(axis=0, dtype='f8')
+    numerator = (weight * (y_true - y_pred) ** 2).sum(axis=0, dtype="f8")
+    denominator = (
+        weight * (y_true - da.average(y_true, axis=0, weights=sample_weight)) ** 2
+    ).sum(axis=0, dtype="f8")
 
     nonzero_denominator = denominator != 0
     nonzero_numerator = numerator != 0
     valid_score = nonzero_denominator & nonzero_numerator
     output_scores = da.ones([y_true.shape[1]], chunks=y_true.chunks[1])
-    output_scores[valid_score] = 1 - (numerator[valid_score] /
-                                      denominator[valid_score])
+    output_scores[valid_score] = 1 - (numerator[valid_score] / denominator[valid_score])
     output_scores[nonzero_numerator & ~nonzero_denominator] = 0.
 
     if isinstance(multioutput, six.string_types):
-        if multioutput == 'raw_values':
+        if multioutput == "raw_values":
             # return scores individually
             return output_scores.compute() if compute else output_scores
-        elif multioutput == 'uniform_average':
+        elif multioutput == "uniform_average":
             # passing None as weights results is uniform mean
             avg_weights = None
-        elif multioutput == 'variance_weighted':
+        elif multioutput == "variance_weighted":
             avg_weights = denominator
     else:
         avg_weights = multioutput
